Corretto controllo index lista percorsi nelle scene
