--- ClassPrj/Assets/_Game/CameraMiniMap.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 7be068039526d8b4c87215176de3e160
-timeCreated: 1456409461
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
