+++ ClassPrj/Assets/_Game/MiniMap.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a492c13b46a42034bad1a4bacc7d1d83
+folderAsset: yes
+timeCreated: 1456841102
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
