+++ ClassPrj/Assets/_Game/MiniMap/CameraMiniMap.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 56682c65bd7f61b41a81d5462bbf6283
+timeCreated: 1456841152
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
