+++ ClassPrj/Assets/_Game/MiniMap/Materials.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 4ab19801b305adb4bbfaaa3e141ae849
+folderAsset: yes
+timeCreated: 1456847185
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
