+++ ClassPrj/Assets/_Game/MiniMap/Materials/minimapTexture.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39c90428f11d8e1419821e74dbbfbda2
+timeCreated: 1456851681
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
