+++ ClassPrj/Assets/_Game/MiniMap/Quad.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6365afb21d0faea45acd9302af0db58d
+timeCreated: 1456851737
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
