+++ ClassPrj/Assets/_Game/MiniMap/minimapTexture.renderTexture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c93c1f2b0bca5b94d853a91d8b6da5e0
+timeCreated: 1456840676
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
