+++ ClassPrj/Assets/_Game/dataBaseDiGioco.asset
@@ -27,4 +27,4 @@ MonoBehaviour:
   - elementoAmicizia: 01000000010000000100000001000000010000000100000001000000
   - elementoAmicizia: 02000000020000000100000001000000010000000100000001000000
   - elementoAmicizia: 01000000010000000100000001000000010000000100000001000000
-  indexPercorsi: ffffffff010000000100000002000000030000000400000005000000
+  indexPercorsi: ffffffff010000000100000002000000ffffffff04000000ffffffff
