+++ ClassPrj/Assets/_Game/dataBasePercorso.asset
@@ -11,10 +11,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2944e3484e1c545478c8085848587dbb, type: 3}
   m_Name: dataBasePercorso
   m_EditorClassIdentifier: 
-  indexPercorsi: 0100000002000000030000000400000005000000
+  indexPercorsi: 010000000200000004000000
   nomePercorsi:
   - Via delle Ninfee
   - Corso dei Pini
-  - Labirinto infernale
   - Via della Palude
-  - Giro della vita
