+++ ClassPrj/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 5.3.3p1
+m_EditorVersion: 5.3.2p4
 m_StandardAssetsVersion: 0
